+++ src/app/terra-components.module.ts
@@ -95,6 +95,7 @@ import { TerraDynamicFormService } from './components/forms/dynamic-form/service
 import { TerraFormFieldControlService } from './components/forms/dynamic-form/service/terra-form-field-control.service';
 import { TerraMultiCheckBoxComponent } from './components/forms/multi-check-box/terra-multi-check-box.component';
 import { TerraFileChooserComponent } from './components/buttons/file-chooser/terra-file-chooser.component';
+import { TerraCodeEditorComponent } from './components/editors/code-editor/terra-code-editor.component';
 // ### import of terra-component examples
 import { TerraButtonComponentExample } from './components/buttons/button/example/terra-button.component.example';
 import { TerraAlertComponentExample } from './components/alert/example/terra-alert.component.example';
@@ -134,6 +135,7 @@ import { TerraDataTableContextMenuComponentExample }
 from './components/tables/data-table/context-menu/example/terra-data-table-context-menu.component.example';
 import { TerraNoteEditorComponentExample } from './components/editors/note-editor/example/terra-note-editor.component.example';
 import { TerraNoteComponentExample } from './components/note/example/terra-note.component.example';
+import { TerraCodeEditorComponentExample } from './components/editors/code-editor/example/terra-code-editor.component.example';
 import { TerraNodeTreeComponentExample } from './components/tree/node-tree/example/terra-node-tree.component.example';
 import { TerraMultiCheckBoxComponentExample } from './components/forms/multi-check-box/example/terra-multi-check-box.component.example';
 import { TerraCategoryPickerComponent } from './components/category-picker/terra-category-picker.component';
@@ -142,6 +144,7 @@ import { HttpClientModule } from '@angular/common/http';
 import { TerraMultiSplitViewBreadcrumbsService } from './components/split-view/multi/injectables/terra-multi-split-view-breadcrumbs.service';
 import { RouterModule } from '@angular/router';
 import { TerraTimePickerComponentExample } from './components/forms/input/time-picker/example/terra-time-picker.component.example';
+
 
 function createCompiler(compilerFactory:CompilerFactory):Compiler
 {
@@ -200,6 +203,7 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraButtonWithOptionsComponent,
         FixedHeaderDirective,
         TerraNoteEditorComponent,
+        TerraCodeEditorComponent,
         TerraTimePickerComponent,
         TerraNoteComponent,
         TerraNodeTreeComponent,
@@ -247,6 +251,7 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraFilterComponentExample,
         TerraDataTableContextMenuComponentExample,
         TerraNoteEditorComponentExample,
+        TerraCodeEditorComponentExample,
         TerraNoteComponentExample,
         TerraNodeTreeComponentExample,
         TerraTimePickerComponentExample
@@ -296,11 +301,12 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraFileChooserComponent,
         TerraButtonWithOptionsComponent,
         TerraNoteEditorComponent,
+        TerraCodeEditorComponent,
         TerraTimePickerComponent,
         TerraNoteComponent,
         TerraNodeTreeComponent,
         TerraNodeComponent,
-        TerraCategoryPickerComponent
+        TerraCategoryPickerComponent,
     ],
     exports:         [
         TerraAlertPanelComponent,
@@ -352,6 +358,7 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraButtonWithOptionsComponent,
         FixedHeaderDirective,
         TerraNoteEditorComponent,
+        TerraCodeEditorComponent,
         TerraTimePickerComponent,
         TerraNoteComponent,
         TerraNodeTreeComponent,
@@ -398,6 +405,7 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraFilterComponentExample,
         TerraDataTableContextMenuComponentExample,
         TerraNoteEditorComponentExample,
+        TerraCodeEditorComponentExample,
         TerraNoteComponentExample,
         TerraDynamicFormComponent,
         TerraDynamicSwitchComponent,